third_party/blink/web_tests/platform/generic/fast/dom/shadow/base-in-shadow-tree-expected.txt

PASS anchorInShadow.href.indexOf("http://www.example.com") is -1
PASS anchorInHostChildren.href.indexOf("http://www.example.com") is -1
PASS anchorOutsideOfShadow.href.indexOf("http://www.example.com") is -1
PASS successfullyParsed is true

TEST COMPLETE

tehepero
